+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 3.8.0 - Unreleased
 - Added a `Create Brotli File` setting to the File Storage driver ([#297](https://github.com/putyourlightson/craft-blitz/issues/297)).
+- Added the ability to use `*` as a wildcard in the `Refresh Cached URLs` utility and `refresh-urls` console command ([#300](https://github.com/putyourlightson/craft-blitz/issues/300)).
 
 ### Changed
 - The values in the `excludedQueryStringParams` are now compared to query string params using regular expressions ([#272](https://github.com/putyourlightson/craft-blitz/issues/272)).